requirements.txt
alembic==1.5.8
Babel==2.9.0
bidict==0.21.2
blinker==1.4
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
cycler==0.10.0
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
Flask-Table==0.5.0
Flask-WTF==0.14.3
greenlet==1.0.0
idna==3.1
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.4
networkx==2.5
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
pycparser==2.20
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.1
scipy==1.6.1
six==1.15.0
SQLAlchemy==1.3.23
tifffile==2021.3.16
Werkzeug==1.0.1
WTForms==2.3.3
